backend/requirements.txt
fastapi
uvicorn
networkx
tree-sitter
tree-sitter-java
sentence-transformers
faiss-cpu
scikit-learn
python-multipart
numpy
python-louvain
